pyproject.toml
[project]
name = "Brotato-ArchipelagoClient"
version = "0.12.0"
dependencies = ["ruamel-yaml>=0.18.16"]

# Copied verbatim from Archipelago (https://github.com/ArchipelagoMW/Archipelago/blob/main/ruff.toml)
[tool.ruff]
line-length = 120
indent-width = 4
target-version = "py311"

[tool.ruff.lint]
select = [
    "B",
    "C",
    "E",
    "F",
    "W",
    "I",
    "N",
    "Q",
    "UP",
    "RET",
    "RSE",
    "RUF",
    "ISC",
    "PLC",
    "PLE",
    "PLW",
    "T20",
    "PERF",
]
ignore = [
    "B011",    # In AP, the use of assert False is essential because we optimise out these statements for release builds.
    "C901",    # Author disagrees with limiting branch complexity
    "N818",    # Author agrees with this rule, but Core AP violates this and changing it would be a hassle.
    "PLC0415", # In AP, we consider local imports totally fine & necessary
    "PLC1802", # Author agrees with this rule, but it literally changes the functionality of the code, which is unsafe.
    "PLC1901", # This is just not equivalent
    "PLE1141", # Gives false positives when the dict keys are tuples, but does not mention this in the suggested fix.
    "UP015",   # Explicit is better than implicit, so we'd prefer to keep "r" in open() calls.
]
